Combined frontend API requests into service object. #5
